better variable types for pxboot

+++ src/seq_sample.cpp
@@ -13,9 +13,9 @@
 
 
 SequenceSampler::SequenceSampler (std::istream* pios, const long int& seed,
-        const double& jackfract, std::string& partf):num_taxa_(0), num_char_(0),
+        const double& jackfract, std::string& partf):num_taxa_(0u), num_char_(0u),
         jkfract_(jackfract), jackknife_(false), partitioned_(false),
-        num_partitioned_sites_(0), num_partitions_(0) {
+        num_partitioned_sites_(0u), num_partitions_(0u) {
     if (seed == -1) {
         srand(get_clock_seed());
     } else {
@@ -35,8 +35,8 @@ SequenceSampler::SequenceSampler (std::istream* pios, const long int& seed,
 void SequenceSampler::read_in_sequences (std::istream* pios) {
     std::string alphaName; // not used, but required by reader
     seqs_ = ingest_alignment(pios, alphaName);
-    num_taxa_ = static_cast<int>(seqs_.size());
-    num_char_ = static_cast<int>(seqs_[0].get_length());
+    num_taxa_ = static_cast<unsigned int>(seqs_.size());
+    num_char_ = static_cast<unsigned int>(seqs_[0].get_length());
     
     // check that it is aligned (doesn't make sense otherwise)
     if (!is_aligned(seqs_)) {
@@ -59,7 +59,7 @@ void SequenceSampler::read_in_sequences (std::istream* pios) {
 
 
 // not currently used
-std::vector<int> SequenceSampler::get_sampled_sites () const {
+std::vector<unsigned int> SequenceSampler::get_sampled_sites () const {
     return sample_sites_;
 }
 
@@ -78,7 +78,7 @@ std::string SequenceSampler::get_resampled_seq (const std::string& origseq) {
 
 
 // this is done once, to populate site sample vector
-void SequenceSampler::sample_sites (const int& numchar) {
+void SequenceSampler::sample_sites (const unsigned int& numchar) {
     if (partitioned_) {
         sample_sites_ = get_partitioned_bootstrap_sites();
     } else if (!jackknife_) {
@@ -90,11 +90,11 @@ void SequenceSampler::sample_sites (const int& numchar) {
 
 
 // sample with replacement.
-std::vector<int> SequenceSampler::get_bootstrap_sites (const int& numchar) {
+std::vector<unsigned int> SequenceSampler::get_bootstrap_sites (const unsigned int& numchar) {
     // numchar zero-initialized elements
-    std::vector<int> randsites (static_cast<unsigned long>(numchar));
+    std::vector<unsigned int> randsites (static_cast<unsigned long>(numchar));
     
-    for (int i = 0; i < numchar; i++) {
+    for (unsigned int i = 0; i < numchar; i++) {
         int randnum = random_int_range(0, (numchar - 1));
         randsites[static_cast<unsigned long>(i)] = randnum;
     }
@@ -105,12 +105,12 @@ std::vector<int> SequenceSampler::get_bootstrap_sites (const int& numchar) {
 
 
 // set up so same composition as original
-std::vector<int> SequenceSampler::get_partitioned_bootstrap_sites () {
-    std::vector<int> master(static_cast<unsigned long>(num_partitioned_sites_), 0);
+std::vector<unsigned int> SequenceSampler::get_partitioned_bootstrap_sites () {
+    std::vector<unsigned int> master(static_cast<unsigned long>(num_partitioned_sites_), 0);
     for (unsigned long i = 0; i < static_cast<unsigned long>(num_partitions_); i++) {
-        auto curNum = static_cast<int>(partitions_[i].size());
+        auto curNum = static_cast<unsigned int>(partitions_[i].size());
         //std::cout << "Partition #" << i << " contains " << curNum << " sites." << std::endl;
-        std::vector<int> randsites = get_bootstrap_sites(curNum);
+        std::vector<unsigned int> randsites = get_bootstrap_sites(curNum);
         for (unsigned long j = 0; j < static_cast<unsigned long>(curNum); j ++) {
         // put partitions back in same spot as original, so partition file does not need to change
             master[static_cast<unsigned long>(partitions_[i][j])] = partitions_[i][randsites[j]];
@@ -120,9 +120,9 @@ std::vector<int> SequenceSampler::get_partitioned_bootstrap_sites () {
 }
 
 // sample WITHOUT replacement. not with partitioned models
-std::vector<int> SequenceSampler::get_jackknife_sites (const int& numchar) {
-    auto numsample = static_cast<int>(numchar * jkfract_ + 0.5);
-    std::vector<int> randsites = sample_without_replacement(numchar, numsample);
+std::vector<unsigned int> SequenceSampler::get_jackknife_sites (const unsigned int& numchar) {
+    auto numsample = static_cast<unsigned int>(numchar * jkfract_ + 0.5);
+    std::vector<unsigned int> randsites = sample_without_replacement(numchar, numsample);
     return randsites;
 }
 
@@ -137,7 +137,7 @@ TODO: update to 1) RAxML and 2) Nexus style
  * So, thrown out first token, and should work for both
 */
 void SequenceSampler::parse_partitions (std::string& partf) {
-    std::vector<int> temp;
+    std::vector<unsigned int> temp;
     std::string line;
     std::ifstream infile(partf.c_str());
     
@@ -151,7 +151,7 @@ void SequenceSampler::parse_partitions (std::string& partf) {
     }
     infile.close();
     
-    num_partitions_ = static_cast<int>(partitions_.size());
+    num_partitions_ = static_cast<unsigned int>(partitions_.size());
     //std::cout << "Found " << num_partitions_ << " partitions." << std::endl;
     calculate_num_partitioned_sites();
     
@@ -163,18 +163,18 @@ void SequenceSampler::parse_partitions (std::string& partf) {
 // expecting pattern: (CHARSET/DNA,) name = start-end[\3][,]
 // want to be flexible with spaces/lackthereof
 // guaranteed to be ordered
-std::vector<int> SequenceSampler::get_partition_sites (const std::string& part) {
-    std::vector<int> sites;
+std::vector<unsigned int> SequenceSampler::get_partition_sites (const std::string& part) {
+    std::vector<unsigned int> sites;
     std::vector<std::string> tokens;
-    int start = 0;
-    int stop = 0;
-    int interval = 1;
+    unsigned int start = 0;
+    unsigned int stop = 0;
+    unsigned int interval = 1;
     
     std::string delim(" -=;\t\\");
     tokenize(part, tokens, delim);
-    get_partition_parameters (tokens, start, stop, interval);
+    get_partition_parameters(tokens, start, stop, interval);
     
-    int i = start;
+    unsigned int i = start;
     while (i <= stop) {
         sites.push_back(i);
         i += interval;
@@ -188,10 +188,10 @@ std::vector<int> SequenceSampler::get_partition_sites (const std::string& part) 
 // not currently used
 /*
 void SequenceSampler::get_site_partitions () {
-    std::vector<int> sites(static_cast<unsigned long>(num_partitioned_sites_), 0);
+    std::vector<unsigned int> sites(static_cast<unsigned long>(num_partitioned_sites_), 0);
     for (unsigned long i = 0; i < static_cast<unsigned long>(num_partitions_); i++) {
         for (unsigned long j = 0; j < partitions_[i].size(); j++) {
-            sites[partitions_[i][j]] = static_cast<int>(i);
+            sites[partitions_[i][j]] = static_cast<unsigned int>(i);
         }
     }
     site_partitions_ = sites;
@@ -203,27 +203,27 @@ void SequenceSampler::get_site_partitions () {
 // after being tokenized, should be of length 4 or 5 (latter when interval)
 // convert from 1-start to 0-start
 void SequenceSampler::get_partition_parameters (std::vector<std::string>& tokens,
-        int& start, int& stop, int& interval) {
-    auto tsize = static_cast<int>(tokens.size());
-    if (tsize < 4 || tsize > 5) {
+        unsigned int& start, unsigned int& stop, unsigned int& interval) {
+    auto tsize = static_cast<unsigned int>(tokens.size());
+    if (tsize < 4u || tsize > 5u) {
         std::cerr << "Error: invalid/unsupported partition specification. Exiting." << std::endl;
         exit(0);
     }
     
-    // ignore first token. will be either CHARSET of DNA,
+    // ignore first token. will be either CHARSET or DNA,
     partition_names_.push_back(tokens[1]);
     //std::cout << "Processing partition '" << tokens[1] << "': ";
     
     if (is_number(tokens[2])) {
-        start = std::atoi(tokens[2].c_str()) - 1;
+        start = std::stoul(tokens[2].c_str()) - 1;
         //std::cout << "start = " << start;
     }
     if (is_number(tokens[3])) {
-        stop = std::atoi(tokens[3].c_str()) - 1;
+        stop = std::stoul(tokens[3].c_str()) - 1;
         //std::cout << "; stop = " << stop;
     }
     if ((static_cast<int>(tokens.size()) == 5) && is_number(tokens[4])) {
-        interval = std::atoi(tokens[4].c_str());
+        interval = std::stoul(tokens[4].c_str());
         //std::cout << "; interval = " << interval;
     }
     //std::cout << std::endl;
@@ -232,8 +232,8 @@ void SequenceSampler::get_partition_parameters (std::vector<std::string>& tokens
 
 void SequenceSampler::calculate_num_partitioned_sites () {
     num_partitioned_sites_ = 0;
-    for (int i = 0; i < num_partitions_; i++) {
-        num_partitioned_sites_ += static_cast<int>(partitions_[static_cast<unsigned long>(i)].size());
+    for (unsigned int i = 0; i < num_partitions_; i++) {
+        num_partitioned_sites_ += static_cast<unsigned int>(partitions_[static_cast<unsigned long>(i)].size());
 //         std::cout << "Partition #" << i << " contains " << (int)partitions[i].size()
 //        << " sites:" << std::endl;
 //         for (unsigned int j = 0; j < partitions[i].size(); j++) {
@@ -244,22 +244,22 @@ void SequenceSampler::calculate_num_partitioned_sites () {
 }
 
 
-int SequenceSampler::get_num_partitioned_sites () const {
+unsigned int SequenceSampler::get_num_partitioned_sites () const {
     return num_partitioned_sites_;
 }
 
 
 // should do some error-checking e.g. for 1) missing sites, 2) overlapping partitions
 void SequenceSampler::check_valid_partitions () {
-    std::vector<int> allSites = partitions_[0];
+    std::vector<unsigned int> allSites = partitions_[0];
     for (unsigned long i = 1; i < static_cast<unsigned long>(num_partitions_); i++) {
         allSites.insert(allSites.end(), partitions_[i].begin(), partitions_[i].end());
     }
     sort(allSites.begin(), allSites.end());
     
-    int max = allSites.back();
-    auto count = static_cast<int>(allSites.size());
-    int diff = max - count + 1;
+    unsigned int max = allSites.back();
+    auto count = static_cast<unsigned int>(allSites.size());
+    unsigned int diff = max - count + 1;
     
     if (diff != 0) { // sites are duplicated
         //std::cerr << "Error in partitioning: maximum site value " << max
@@ -270,19 +270,19 @@ void SequenceSampler::check_valid_partitions () {
 
 
 // find sites 1) present in more than 1 partition or 2) not present in any
-[[ noreturn ]] void SequenceSampler::find_duplicates_missing (const std::vector<int>& allSites) {
-    std::vector<int> unique;
-    std::vector<int> duplicates;
-    std::vector<int> missing;
+[[ noreturn ]] void SequenceSampler::find_duplicates_missing (const std::vector<unsigned int>& allSites) {
+    std::vector<unsigned int> unique;
+    std::vector<unsigned int> duplicates;
+    std::vector<unsigned int> missing;
     
-    int maxVal = static_cast<int>(allSites.back());
+    unsigned int maxVal = static_cast<unsigned int>(allSites.back());
     
-    std::vector<int> counts(static_cast<unsigned long>(maxVal), 0);
-    for (int allSite : allSites) {
+    std::vector<unsigned int> counts(static_cast<unsigned long>(maxVal), 0);
+    for (unsigned int allSite : allSites) {
         counts[static_cast<unsigned long>(allSite)]++;
     }
     
-    for (int i = 0; i < maxVal; i++) {
+    for (unsigned int i = 0; i < maxVal; i++) {
         switch (counts[static_cast<unsigned long>(i)]) {
             case 0:
                 missing.push_back(i);
@@ -298,7 +298,7 @@ void SequenceSampler::check_valid_partitions () {
     if (!duplicates.empty()) {
         std::cerr << "Error: the following " << duplicates.size()
                 << " sites are found in more than one partition: ";
-        for (int duplicate : duplicates) {
+        for (unsigned int duplicate : duplicates) {
             std::cerr << duplicate << " ";
         }
         std::cerr << std::endl;
@@ -306,7 +306,7 @@ void SequenceSampler::check_valid_partitions () {
     if (!missing.empty()) {
         std::cerr << "Error: the following " << missing.size()
                 << " sites are not found in any partition: ";
-        for (int ms : missing) {
+        for (unsigned int ms : missing) {
             std::cerr << ms << " ";
         }
         std::cerr << std::endl;
